chore: Update npm dependency to latest stable version and add Jest as a new dependency
This commit updates the npm dependency in the package.json file to the latest stable version and adds Jest as a new dependency. Jest is a popular JavaScript testing framework that will be used for unit testing in the project. This update ensures that the project is using the latest stable version of the dependency and includes the necessary testing framework for future development.